requirements.txt
mysql-connector-python==8.0.19
psycopg2-binary==2.8.4
